migrate check_library_usage_telemetry to _common (#54355)
Fixes: https://github.com/ray-project/ray/issues/53478

- migrating check_library_usage_telemetry from `_private` to `_common`
- migrating TelemetryCallsite from `_private` to `_common`.
