+++ python/ray/train/v2/tests/test_telemetry.py
@@ -4,7 +4,7 @@ import pytest
 
 import ray
 import ray._private.usage.usage_lib as ray_usage_lib
-from ray._private.test_utils import TelemetryCallsite, check_library_usage_telemetry
+from ray._common.test_utils import TelemetryCallsite, check_library_usage_telemetry
 from ray.train.v2.api.data_parallel_trainer import DataParallelTrainer
